Separate crate for aptos-evm
