--- aptos-move/aptos-vm/src/evm/mod.rs
@@ -1,8 +1,0 @@
-// Copyright © Aptos Foundation
-
-mod eth_address;
-#[cfg(test)]
-mod in_memory_storage;
-mod utils;
-mod evm_backend;
-mod engine;
