+++ aptos-move/framework/aptos-framework/doc/evm.md
@@ -5,57 +5,9 @@
 
 
 
--  [Function `create`](#0x1_evm_create)
--  [Function `call`](#0x1_evm_call)
 
 
 <pre><code></code></pre>
-
-
-
-<a name="0x1_evm_create"></a>
-
-## Function `create`
-
-
-
-<pre><code><b>public</b> <b>fun</b> <a href="evm.md#0x1_evm_create">create</a>()
-</code></pre>
-
-
-
-<details>
-<summary>Implementation</summary>
-
-
-<pre><code><b>public</b> <b>native</b> <b>fun</b> <a href="evm.md#0x1_evm_create">create</a>();
-</code></pre>
-
-
-
-</details>
-
-<a name="0x1_evm_call"></a>
-
-## Function `call`
-
-
-
-<pre><code><b>public</b> <b>fun</b> <a href="evm.md#0x1_evm_call">call</a>()
-</code></pre>
-
-
-
-<details>
-<summary>Implementation</summary>
-
-
-<pre><code><b>public</b> <b>native</b> <b>fun</b> <a href="evm.md#0x1_evm_call">call</a>();
-</code></pre>
-
-
-
-</details>
 
 
 [move-book]: https://aptos.dev/move/book/SUMMARY
