+++ aptos-move/framework/aptos-framework/sources/evm/evm.move
@@ -1,7 +1,7 @@
 
 module aptos_framework::evm {
-
-    public native fun create(caller: H160, value: u256, init_code: Vec<u8>, gas_limit: u64);
-
-    public native fun call(caller: H160, address: H160, value: u256, data: Vec<u8>, gas_limit: u64);
+    //
+    // public native fun create(caller: H160, value: u256, init_code: Vec<u8>, gas_limit: u64);
+    //
+    // public native fun call(caller: H160, address: H160, value: u256, data: Vec<u8>, gas_limit: u64);
 }
